requirements.txt
google-cloud-vision
spacy
